libs/shared_kafka/requirements.txt
kafka-python==2.1.5
pytest==8.3.5
